Add `wigos_station_identifier` to WISNotification
Split messages into their own files
